refactor(sampling): clarify packed policy ownership
Move batch policy declarations out of the executor interface, resolve complete fixed-fault worker budgets once, and remove dead batch indirections and accessors.
